+++ sass/main.scss
@@ -7,5 +7,6 @@
 @import 'modules/_logInPage';
 @import 'modules/_dashboard';
 @import 'modules/learnPage';
+@import 'modules/_widget';
 @import 'modules/aboutPage';
