+++ sass/modules/_widget.scss
@@ -1,7 +1,7 @@
 #map {
     position: relative;
     margin: 0 auto;
-    max-width: 80%;
+    max-width: 70%;
     height: 350px;
     #info-wrp {
         position: absolute;
